Prepare OECD Data on Labour Force Participation Rates

Process labour force participation data downloaded from the OECD Data Explorer.

Usage

data_ssvd_lfp(data, age_max = 75, n_comp = 5)

Arguments

data

A data frame containing OECD data.

age_max

The upper limit of the oldest closed age group. Default is 75.

n_comp

Number of SVD components to include in result. Default is 5.

Value

A tibble with the format required by function bage::ssvd().

Usage

Step 1: Download data

library(rsdmx)
url <- paste("https://sdmx.oecd.org/public/rest/data",
             "OECD.ELS.SAE,DSD_LFS@DF_LFS_INDIC,1.1",
             "all?dimensionAtObservation=AllDimensions",
             sep = "/")
lfp_sdmx <- rsdmx::readSDMX(url)    ## can be slow
data <- as.data.frame(lfp_sdmx)     ## can also be slow

Step 2: Call function 'data_ssvd_lfp()'

lfp_data <- data_ssvd_lfp(data)

See also

Examples

lfp_data_small <- data_ssvd_lfp(oecd_lfp)
#> Tidying data...
#> Carrying out SVD for 'total'...
#> Carrying out SVD for 'indep'...
#> Carrying out SVD for 'joint'...
#> Combining results...